Just pass the pointer in and out

+++ j4rs/src/main/java/org/clulab/transformers/tokenizer/j4rs/JavaJ4rsTokenizer.java
@@ -9,25 +9,25 @@ public class JavaJ4rsTokenizer {
     private static native Instance create(Instance<String> name);
 
     // Garbage collect the RustTokenizer.
-    public static native void destroy(Instance<Integer> rustTokenizerId);
+    public static native void destroy(Instance<Long> rustTokenizerId);
 
     // Perform tokenization on the words.
-    private static native Instance tokenize(Instance<Integer> rustTokenizerId, Instance<String[]> words);
+    private static native Instance tokenize(Instance<Long> rustTokenizerId, Instance<String[]> words);
 
-    public static int create(String name) {
+    public static long create(String name) {
         Instance name_instance = Java2RustUtils.createInstance(name);
         Instance tokenizer_id_instance = create(name_instance);
-        int tokenizer_id = Java2RustUtils.getObjectCasted(tokenizer_id_instance);
+        long tokenizer_id = Java2RustUtils.getObjectCasted(tokenizer_id_instance);
 
         return tokenizer_id;
     }
 
-    public static void destroy(int tokenizer_id) {
+    public static void destroy(long tokenizer_id) {
         Instance tokenizer_id_instance = Java2RustUtils.createInstance(tokenizer_id);
         destroy(tokenizer_id_instance);
     };
 
-    public static JavaJ4rsTokenization tokenize(int tokenizer_id, String[] words) {
+    public static JavaJ4rsTokenization tokenize(long tokenizer_id, String[] words) {
         Instance tokenizer_id_instance = Java2RustUtils.createInstance(tokenizer_id);
         Instance words_instance = Java2RustUtils.createInstance(words);
         Instance tokenization_instance = tokenize(tokenizer_id_instance, words_instance);
